# backend/api/src/main/resources/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kalico API",
    "description": "REST API for Kalico backend services",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.kalico.ai",
      "description": "Prod"
    },
    {
      "url": "http://localhost:8080",
      "description": "Dev"
    }
  ],
  "paths": {
...
    "/cms/posts/all": {
      "get": {
        "tags": [
          "cms"
        ],
        "summary": "Get list of posts",
        "description": "Get a list of posts",
        "operationId": "getAllPosts",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All posts response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageableRecipeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cms/posts/{id}": {
      "get": {
        "tags": [
          "cms"
        ],
        "summary": "Get a post by id",
        "description": "Get a post by id",
        "operationId": "getPostById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get single post response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeFull"
                }
              }
            }
          }
        }
      }
    },
    "/cms/post/update": {
      "post": {
        "tags": [
          "cms"
        ],
        "summary": "Update post",
        "description":"Update post",
        "operationId": "updatePost",
        "requestBody": {
          "required": true,
          "description": "Update content request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeFull"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update content response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GenericResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PageableRecipeResponse": {
        "type": "object",
        "required": [
          "total_records",
          "num_pages",
          "records"
        ],
        "properties": {
          "total_records": {
            "type": "integer",
            "format": "int64"
          },
          "num_pages": {
            "type": "integer",
            "format": "int32"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeLite"
            }
          }
        }
      },
      "RecipeLite": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "published": {
            "type": "boolean"
          },
          "slug": {
            "type": "string"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "num_ingredients": {
            "type": "integer",
            "format": "int32"
          },
          "num_steps": {
            "type": "integer",
            "format": "int32"
          },
          "cooking_time": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RecipeStep": {
        "type": "object",
        "properties": {
          "step_number": {
            "type": "integer",
            "format": "int32"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "text": {
            "type": "string"
          }
        }
      },
      "Ingredient": {
        "type": "object",
        "nullable": true,
        "properties": {
          "ingredient": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "units": {
            "type": "string"
          }
        }
      },
      "RecipeFull": {
        "type": "object",
        "properties": {
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            }
          },
          "instructions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeStep"
            }
          },
          "source": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "keywords": {
            "type": "string"
          },
          "recipe_lite": {
            "$ref": "#/components/schemas/RecipeLite"
          },
          "prev": {
            "$ref": "#/components/schemas/RecipeLite"
          },
          "next": {
            "$ref": "#/components/schemas/RecipeLite"
          }
        }
      }
    }
  }
}